Temporarily revert changes to def/es6.js to publish v0.4.13.
I was meaning to bump the minor version to v0.5 because of @caridy's
backwards-incompatible changes to ES6 module definitions (#112), but I
accidentally published them as v0.4.12.

I'm going to publish one final version on the v0.4 line so that code
depending on that minor version won't break, and then revert this commit
and publish that as v0.5.0.