Fix named function expression scopes.
The commit that added support for function expression names counting as
part of the binding for that function’s scope,
94adb160509d35a4c03bccd49030296770737e7a, caused only the function name
to count as a binding in that function. For example:

    var f = function f(a, b) { var c; }

Inside `f`, the only bound name was `f` itself referring to the
function expression. It should have bound `a`, `b`, `c`, and `f`.