fix: type field type
This is causing me issues when introspecting the defined types for
typescript definition file generation. Buildable types have their
`type` field automatically set, but I believe it's being done
incorrectly. For example `Identifier` has it's `type` field set to
type "Identifier", with a default value of "Identifier". It should
be set to type "string" with a default value of "Identifier".